don't allow notes without an associated clue
in a puzzle with unchecked squares, the cursor
can be in a state where there is no associated
clue (if the current square is only in an across
and the filldir is down, or vice versa). if a
player tries to add a note in that state, print
a suggestion to change the filldir instead of
starting the note prompt.