always include current clue in clue list
previously, the draw_clues() function never
drew more than the specified max number of
lines. for most puzzles this was fine, but
it sometimes stopped before drawing the
current clue, e.g. in small puzzles with
long clues. this adds a check that we've
drawn the current clue before finishing.

this isn't perfect -- if some across clues
are long enough, then the across cluebox
can overlap the down cluebox. a better
solution might be to build up the cluebox
around the current clue, alternating
earlier and later clues until the max
number of lines.